change: extracted formatTime
