+++ src/functions/format/formatTime.test.ts
@@ -1,0 +1,4 @@
+import { describe } from "vitest";
+
+// TODO: Write tests
+describe("formatTime", () => {});
